# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SubastApp API",
    "description": "API documentation for SubastApp",
    "license": {
      "name": "FabrizioB Website",
      "url": "https://fabriziob.com/"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4002",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Listar productos",
        "description": "Devuelve todos los productos con sus datos y lista de IDs de fotos",
        "operationId": "listar",
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductoResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "summary": "Crear producto",
        "description": "Crea un producto con sus datos y opcionalmente una lista de imágenes",
        "operationId": "crear",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "datos": {
                    "$ref": "#/components/schemas/ProductoRequest"
                  },
                  "imagenes": {
                    "type": "array",
                    "description": "Lista de imágenes del producto (PNG, JPEG, WEBP)",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "datos"
                ]
              },
              "encoding": {
                "datos": {
                  "contentType": "application/json"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
    "/productos/{id}/fotos": {
      "post": {
        "tags": [
          "Productos"
        ],
        "summary": "Subir fotos",
        "description": "Agrega una o más imágenes a un producto existente. Devuelve la lista de IDs generados",
        "operationId": "subirFotos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imagenes": {
                    "type": "array",
                    "description": "Lista de imágenes (PNG, JPEG, WEBP)",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "imagenes"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fotos subidas, devuelve lista de IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "example": "42"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "400": {
            "description": "No se enviaron imágenes"
          }
        }
      }
    },
...
    "/productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener producto por ID",
        "description": "Devuelve un producto con sus datos completos, extra y lista de IDs de fotos",
        "operationId": "obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar producto",
        "description": "Actualiza parcialmente los campos de un producto. Solo para uso interno de la empresa",
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Campos a actualizar, solo se pisan los que vienen en el body",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
    "/productos/{id}/habilitacion": {
      "patch": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar habilitación del producto",
        "description": "Cambia el estado del bien. Si el estado es 'rechazado', motivoRechazo es obligatorio. Solo para uso interno de la empresa",
        "operationId": "patchEstado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Nuevo estado del bien",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoHabilitacionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Estado actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoResponse"
                }
              }
            }
          },
          "400": {
            "description": "motivoRechazo requerido al rechazar"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
...
    "/productos/{id}/seguro": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener seguro del producto",
        "description": "Devuelve el seguro completo si la póliza existe en la tabla de seguros, solo el número si está registrado pero no existe en la tabla, o indica que no tiene seguro",
        "operationId": "seguro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Seguro encontrado o sin seguro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoSeguroResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
    "/productos/{id}/fotos/{imgId}/content": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener imagen",
        "description": "Devuelve el contenido binario de una foto. El Content-Type se detecta automáticamente (PNG, JPEG o WEBP)",
        "operationId": "fotoContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          },
          {
            "name": "imgId",
            "in": "path",
            "description": "ID de la imagen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 12
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen encontrada",
            "content": {
              "image/png": {},
              "image/jpeg": {},
              "image/webp": {}
            }
          },
          "404": {
            "description": "Foto o producto no encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductoRequest": {
        "type": "object",
        "description": "Datos del producto en JSON",
        "properties": {
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "disponible": {
            "type": "string"
          },
          "descripcionCatalogo": {
            "type": "string"
          },
          "descripcionCompleta": {
            "type": "string"
          },
          "declaracionPropiedad": {
            "type": "boolean"
          },
          "esPiezaMultiple": {
            "type": "boolean"
          },
          "cantidadPiezas": {
            "type": "integer",
            "format": "int32"
          },
          "esObraDeArte": {
            "type": "boolean"
          },
          "artista": {
            "type": "string"
          },
          "fechaCreacionObra": {
            "type": "string",
            "format": "date"
          },
          "historia": {
            "type": "string"
          },
          "deposito": {
            "type": "string"
          }
        }
      },
      "EstadoBien": {
        "type": "string",
        "enum": [
          "recibido",
          "inspeccionado",
          "aceptado",
          "rechazado",
          "en_subasta",
          "vendido",
          "devuelto"
        ]
      },
      "ProductoResponse": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "disponible": {
            "type": "string"
          },
          "descripcionCatalogo": {
            "type": "string"
          },
          "descripcionCompleta": {
            "type": "string"
          },
          "revisor": {
            "type": "integer",
            "format": "int32"
          },
          "duenio": {
            "type": "integer",
            "format": "int32"
          },
          "seguro": {
            "type": "string"
          },
          "fotosIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "estadoBien": {
            "$ref": "#/components/schemas/EstadoBien"
          },
          "motivoRechazo": {
            "type": "string"
          },
          "declaracionPropiedad": {
            "type": "boolean"
          },
          "esPiezaMultiple": {
            "type": "boolean"
          },
          "cantidadPiezas": {
            "type": "integer",
            "format": "int32"
          },
          "esObraDeArte": {
            "type": "boolean"
          },
          "artista": {
            "type": "string"
          },
          "fechaCreacionObra": {
            "type": "string",
            "format": "date"
          },
          "historia": {
            "type": "string"
          },
          "deposito": {
            "type": "string"
          }
        }
      },
...
      "ProductoUpdateRequest": {
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "disponible": {
            "type": "string"
          },
          "descripcionCatalogo": {
            "type": "string"
          },
          "descripcionCompleta": {
            "type": "string"
          },
          "revisor": {
            "type": "integer",
            "format": "int32"
          },
          "declaracionPropiedad": {
            "type": "boolean"
          },
          "esPiezaMultiple": {
            "type": "boolean"
          },
          "cantidadPiezas": {
            "type": "integer",
            "format": "int32"
          },
          "esObraDeArte": {
            "type": "boolean"
          },
          "artista": {
            "type": "string"
          },
          "fechaCreacionObra": {
            "type": "string",
            "format": "date"
          },
          "historia": {
            "type": "string"
          },
          "deposito": {
            "type": "string"
          }
        }
      },
      "ProductoHabilitacionUpdateRequest": {
        "type": "object",
        "properties": {
          "estadoBien": {
            "$ref": "#/components/schemas/EstadoBien"
          },
          "motivoRechazo": {
            "type": "string"
          },
          "revisor": {
            "type": "integer",
            "format": "int32"
          },
          "seguro": {
            "type": "string"
          }
        },
        "required": [
          "estadoBien"
        ]
      },
...
      "PageProductoResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductoResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
      "ProductoSeguroResponse": {
        "type": "object",
        "properties": {
          "seguroObj": {
            "$ref": "#/components/schemas/Seguro"
          },
          "seguro": {
            "type": "string"
          }
        }
      },
      "Seguro": {
        "type": "object",
        "properties": {
          "nroPoliza": {
            "type": "string"
          },
          "compania": {
            "type": "string"
          },
          "polizaCombinada": {
            "type": "string"
          },
          "importe": {
            "type": "number"
          }
        }
      },
...
    }
  }
}